Redesign Optional Comparison
